LayoutTests/http/tests/media/media-source/video-media-source-add-and-remove-ids.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/media-resources/video-test.js"></script>
        <script src="/media/resources/media-source/webm/segment-info.js"></script>
        <script src="media-source.js"></script>
        <script>
            var segmentHelper = new MediaSourceTest.SegmentHelper(WebMSegmentInfo.testWebM);
            var defaultSourceMimetype = segmentHelper.segmentInfo.type;

            function expectExceptionOnAddId(videoTag, id, type, error)
            {
                try {
                    videoTag.webkitSourceAddId(id, type);
                    failTest("Expected an exception");
                } catch (e) {
                    if (!(e.code == error)) {
                        failTest("Unexpected exception " + e);
                        throw e;
                    }
                    consoleWrite("Got expected exception " + e);
                }
            }

            function expectExceptionOnRemoveId(videoTag, id, error)
            {
                try {
                    videoTag.webkitSourceRemoveId(id);
                    failTest("Expected an exception");
                } catch (e) {
                    if (!(e.code == error)) {
                        failTest("Unexpected exception " + e);
                        throw e;
                    }
                    consoleWrite("Got expected exception " + e);
                }
            }

            function expectExceptionOnAppend(videoTag, id, buf, error)
            {
                try {
                    videoTag.webkitSourceAppend(id, buf);
                    failTest("Expected an exception");
                } catch (e) {
                    if (!(e.code == error)) {
                        failTest("Unexpected exception " + e);
                        throw e;
                    }
                    consoleWrite("Got expected exception " + e);
                }
            }

            function testAddIdWhileClosed(videoTag)
            {
                consoleWrite("Test adding an ID while closed.");
                expectExceptionOnAddId(videoTag, "123", defaultSourceMimetype, DOMException.INVALID_STATE_ERR);
            }

            function testAddIdFailureCases(runNextTestCase, videoTag)
            {
                consoleWrite("Test empty ID case");
                expectExceptionOnAddId(videoTag, "", defaultSourceMimetype, DOMException.INVALID_ACCESS_ERR);

                videoTag.webkitSourceAddId("123", defaultSourceMimetype);

                consoleWrite("Test adding the same ID again.");
                expectExceptionOnAddId(videoTag, "123", defaultSourceMimetype, DOMException.INVALID_STATE_ERR);
                videoTag.webkitSourceRemoveId("123");

                consoleWrite("Test empty type.");
                expectExceptionOnAddId(videoTag, "234", "", DOMException.INVALID_ACCESS_ERR);

                consoleWrite("Test an unsupported type.");
                expectExceptionOnAddId(videoTag, "234", "audio/x-unsupported-format", DOMException.NOT_SUPPORTED_ERR);

                consoleWrite("Test a supported type with an unsupported codec.");
                expectExceptionOnAddId(videoTag, "234", "video/webm; codecs=\"vp8, speex\"", DOMException.NOT_SUPPORTED_ERR);

                consoleWrite("Test reaching sourceID limit.");
                var reachedIdLimit = false;

                // The 20 here is an arbitrary upper limit to make sure the test terminates. This test
                // assumes that implementations won't support more than 20 sourceID's simultaneously.
                for (var i = 0; i < 20; ++i) {
                    var sourceID = "sourceID-" + i;
                    try {
                        videoTag.webkitSourceAddId(sourceID, defaultSourceMimetype);
                    } catch(e) {
                        if (e.code != DOMException.QUOTA_EXCEEDED_ERR) {
                            failTest("Unexpected exception " + e);
                            throw e;
                        }
                        reachedIdLimit = true;
                        break;
                    }
                }

                if (!reachedIdLimit) {
                    failTest("Failed to reach SourceID limit.");
                    return;
                }

                consoleWrite("Test that SourceIDs can't be added while in the ended state.");
                videoTag.webkitSourceEndOfStream(videoTag.EOS_NO_ERROR);
                expectExceptionOnAddId(videoTag, "123", defaultSourceMimetype, DOMException.INVALID_STATE_ERR);

                runNextTestCase();
            }

            function testRemoveEmptyId(runNextTestCase, videoTag)
            {
                consoleWrite("Test empty ID case");
                expectExceptionOnRemoveId(videoTag, "", DOMException.INVALID_ACCESS_ERR);

                runNextTestCase();
            }

            function testRemoveNonexistentId(runNextTestCase, videoTag)
            {
                consoleWrite("Test removing an ID that was never added.");
                expectExceptionOnRemoveId(videoTag, "345", DOMException.SYNTAX_ERR);

                runNextTestCase();
            }

            function testRemoveAgain(runNextTestCase, videoTag)
            {
                consoleWrite("Test removing an ID that was already removed.");
                videoTag.webkitSourceAddId("123", defaultSourceMimetype);
                videoTag.webkitSourceRemoveId("123");
                expectExceptionOnRemoveId(videoTag, "123", DOMException.SYNTAX_ERR);

                runNextTestCase();
            }

            function testRemoveIdAfterEnded(runNextTestCase, videoTag)
            {
                consoleWrite("Test that an ID can be removed while in the ended state.");
                videoTag.webkitSourceAddId("123", defaultSourceMimetype);
                videoTag.webkitSourceEndOfStream(videoTag.EOS_NO_ERROR);
                videoTag.webkitSourceRemoveId("123");

                runNextTestCase();
            }

            function testAddIdAfterRemoving(runNextTestCase, videoTag)
            {
                consoleWrite("Test that an ID can be added again after it is removed.");
                videoTag.webkitSourceAddId("123", defaultSourceMimetype);
                videoTag.webkitSourceRemoveId("123");

                try {
                    videoTag.webkitSourceAddId("123", defaultSourceMimetype);
                } catch (e) {
                    consoleWrite("Unexpected exception: " + e);
                }

                runNextTestCase();
            }

            function testAppendFailureCases(runNextTestCase, videoTag)
            {
                var initSegment = segmentHelper.initSegment;
                var mediaSegment = segmentHelper.mediaSegments[0];

                videoTag.webkitSourceAddId("123", defaultSourceMimetype);

                consoleWrite("Test append with empty ID.");
                expectExceptionOnAppend(videoTag, "", initSegment, DOMException.INVALID_ACCESS_ERR);

                consoleWrite("Test append with an invalid ID");
                expectExceptionOnAppend(videoTag, "234", initSegment, DOMException.SYNTAX_ERR);

                consoleWrite("Test append with a null buffer.");
                expectExceptionOnAppend(videoTag, "123", null, DOMException.INVALID_ACCESS_ERR);

                consoleWrite("Test a successful append.");
                videoTag.webkitSourceAppend("123", initSegment);

                videoTag.webkitSourceRemoveId("123");
                expectExceptionOnAppend(videoTag, "123", mediaSegment, DOMException.SYNTAX_ERR);

                runNextTestCase();
            }

            function onLoad()
            {
                findMediaElement();

                waitForEvent('loadstart');
                waitForEvent('webkitsourceopen');

                segmentHelper.init(video, function(success)
                {
                    if (!success) {
                        failTest("Failed to load segment data");
                        return;
                    }

                    testAddIdWhileClosed(video);

                    var testCases = [
                        testAddIdFailureCases,
                        testRemoveEmptyId,
                        testRemoveNonexistentId,
                        testRemoveAgain,
                        testRemoveIdAfterEnded,
                        testAddIdAfterRemoving,
                        testAppendFailureCases,
                    ];

                    MediaSourceTest.startSourceOpenTesting(video, testCases);
                });
            }
        </script>
    </head>
    <body onload="onLoad()">
        <video> </video>
        <p>Tests webkitSourceAddId() &amp; webkitSourceRemoveId() methods</p>
    </body>
</html>
