LayoutTests/http/tests/media/media-source/video-media-source-add-and-remove-ids-expected.txt

Tests webkitSourceAddId() & webkitSourceRemoveId() methods

Test adding an ID while closed.
Got expected exception Error: INVALID_STATE_ERR: DOM Exception 11

running testAddIdFailureCases
EVENT(loadstart)
EVENT(webkitsourceopen)
Test empty ID case
Got expected exception Error: INVALID_ACCESS_ERR: DOM Exception 15
Test adding the same ID again.
Got expected exception Error: INVALID_STATE_ERR: DOM Exception 11
Test empty type.
Got expected exception Error: INVALID_ACCESS_ERR: DOM Exception 15
Test an unsupported type.
Got expected exception Error: NOT_SUPPORTED_ERR: DOM Exception 9
Test a supported type with an unsupported codec.
Got expected exception Error: NOT_SUPPORTED_ERR: DOM Exception 9
Test reaching sourceID limit.
Test that SourceIDs can't be added while in the ended state.
Got expected exception Error: INVALID_STATE_ERR: DOM Exception 11

running testRemoveEmptyId
EVENT(loadstart)
EVENT(webkitsourceopen)
Test empty ID case
Got expected exception Error: INVALID_ACCESS_ERR: DOM Exception 15

running testRemoveNonexistentId
EVENT(loadstart)
EVENT(webkitsourceopen)
Test removing an ID that was never added.
Got expected exception Error: SYNTAX_ERR: DOM Exception 12

running testRemoveAgain
EVENT(loadstart)
EVENT(webkitsourceopen)
Test removing an ID that was already removed.
Got expected exception Error: SYNTAX_ERR: DOM Exception 12

running testRemoveIdAfterEnded
EVENT(loadstart)
EVENT(webkitsourceopen)
Test that an ID can be removed while in the ended state.

running testAddIdAfterRemoving
EVENT(loadstart)
EVENT(webkitsourceopen)
Test that an ID can be added again after it is removed.
Unexpected exception: Error: QUOTA_EXCEEDED_ERR: DOM Exception 22

running testAppendFailureCases
EVENT(loadstart)
EVENT(webkitsourceopen)
Test append with empty ID.
Got expected exception Error: INVALID_ACCESS_ERR: DOM Exception 15
Test append with an invalid ID
Got expected exception Error: SYNTAX_ERR: DOM Exception 12
Test append with a null buffer.
Got expected exception Error: INVALID_ACCESS_ERR: DOM Exception 15
Test a successful append.
Got expected exception Error: SYNTAX_ERR: DOM Exception 12
END OF TEST
